+++ api/core/gee_init.py
@@ -1,0 +1,6 @@
+import ee
+
+# Authenticate and initialize Earth Engine
+def init_gee():
+    ee.Authenticate()
+    ee.Initialize(project="gee-dashboard-project")
